==> step3_output_summary.Rmd <==
---
title: 'null'
output:
  word_document: default
  html_document: default
  pdf_document: default
---
```{r setup}
knitr::opts_knit$set(root.dir = './output')
knitr::opts_chunk$set(echo=FALSE, warning = FALSE, message = FALSE)

```

```{r Initial formatting, include=FALSE}
knitr::opts_chunk$set(echo=FALSE)


load("../FOR_OUTPUT_RMD_v2.RData")

library(nlme)
library(MuMIn)
library(tidyverse)
library(flextable)
library(ggplot2)
library(ggh4x)
devtools::install_github("thomasp85/scico")
library(scico)
library(tidyverse)

#create idealized dataframe with fixed speed for plotting
ideal <- expand.grid(ID=unique(summary_model$ID),
                     temp_mean=c(12:21),
                     press_mean = c(1,8),
                     speedBL = c(0.6)
                  
)

ideal <- ideal %>% 
  left_join(summary_model %>% select(ID, startmass, run, tunnel), by = "ID")

ideal <- ideal %>%  distinct(.keep_all = T)

ideal$resp <- exp(predict(lme.plot, newdata = ideal, group = ID))
ideal$resp_glob <- exp(predict(lme.plot, newdata = ideal, level = "0"))
ideal_1 <- ideal %>% filter(press_mean == 1)

#not used now but possibly useful
min_max <- expand.grid(ID=unique(summary$ID),
                     temp_mean=c(12, 12.7, 19.6, 20.7),
                     press_mean = c(1),
                     speedBL = c(0.6))
min_max$resp_glob <- exp(predict(lme.plot, newdata = min_max, level = "0"))

summary <- summary %>% 
  mutate(speedBL_min = speed_min/(length/100),
         speedBL_max = speed_max/(length/100),
         log_resp = log(resp))

summary_model <- summary_model %>% 
  mutate(speedBL_min = speed_min/(length/100),
         speedBL_max = speed_max/(length/100))

#calculate Q10
Q10at1 <- (ideal$resp_glob[70]/ideal$resp_glob[1])^(10/(ideal$temp_mean[70]-ideal$temp_mean[1]))
Q10at8 <- (ideal$resp_glob[140]/ideal$resp_glob[71])^(10/(ideal$temp_mean[140]-ideal$temp_mean[71]))
Q10 <- rbind(Q10at1, Q10at8)
flextable(as.data.frame(Q10))
```

## Tables for manuscript

**Tab 1** Summary of biological and experimental paramaters by individual eel (ranges refer to means per measurement)

```{r Tab 1}
biomass_backup <-biomasses
#biomasses <- biomass_backup

biomasses$ID <- biomasses$ID %>% str_replace_all(c("a" = "A", "b" = "B", "c" = "C", "d" = "D", "e" = "E", "f" = "F", "g" = "G", "h" = "H", "i" = "I"))

biomasses_temp<- biomasses %>% 
  mutate(ID_old = as.factor(biomasses$ID_old))

biomasses <- biomasses %>% 
  rename(massstart = "Biomass at start (g)",
         massend = "Biomass at end (g)",
         fatstart = "Fat at start (% WM)",
         fatend = "Fat at end (% WM)") %>% 
  mutate(fatstart = round(fatstart, 1),
         fatend = round(fatend, 1),
         "Biomass at start/end (g)" = paste(massstart, massend, sep = "/"),
         "Fat at start/end (%WM)" = paste(fatstart, fatend, sep = "/"),
         massstart = NULL,
         massend =  NULL,
         fatstart = NULL,
         fatend =  NULL)

         
swimtime_temp <- swimtime %>%  
    mutate(ID = as.factor(swimtime$ID))

summ_speed <- summary %>% 
  group_by(ID) %>% 
  summarize(speed_min = min(speedBL_min),
            speed_max = max(speedBL_max),
            min_temp = min(temp_min),
            max_temp = max(temp_max),
            min_press = min(press_min),
            max_press = max(press_max)) %>%
  left_join(biomasses_temp, by = c("ID" = "ID_old")) %>% 
  mutate(ID = ID.y) %>% 
  select(ID, min_press, max_press, min_temp, max_temp, speed_min, speed_max)

align(flextable(biomasses %>%
            left_join((swimtime %>% select(ID, time_d, distance)), by = c("ID_old" = "ID")) %>%
            left_join(summ_speed, by = "ID") %>%
            select("ID", "Length (cm)", "Silvering Index*", "Biomass at start/end (g)", "Fat at start/end (%WM)", "Silvering Index*", "time_d", "distance", "min_press", "max_press", "min_temp", "max_temp", "speed_min", "speed_max") %>%
            mutate(time_d = round(time_d, 1),
                   distance = round(distance, 0),
                   speed_min = round(speed_min, 2),
                   speed_max = round(speed_max, 2),
                   temp_min = round(min_temp, 1),
                   temp_max = round(max_temp, 1),
                   press_min = round(min_press, 2),
                   press_max = round(max_press, 2),
                   "Pressure range (bar)" = paste(min_press, max_press, sep = "-"),
                   "Temperature range (°C)" = paste(min_temp, max_temp, sep = "-"),
                   "Velocity range (BL/s)" = paste(speed_min, speed_max, sep = "-")) %>% 
            rename("Swim time (d)" = time_d,
                   "Distance (km)" = distance) %>% 
              select(-speed_min, -speed_max, -temp_min, -temp_max, -press_min, -press_max, -min_temp,- max_temp, -min_press, -max_press)) %>% 
  colformat_num(big.mark   = "") %>%
  autofit(add_w = -0.5), align = "center", part = "all")
```

\*according to Durif et al. (2005)\
\
\
\

**Tab X** Observed pressure, temperature and velocity ranges during measurements and resulting average oxygen consumption rates (mg/kg * h^-1) (for individual values see Tab S1).

```{r Tab 3}

summ_by_ID <- summary_model %>% 
            group_by(ID, ntemp, npress) %>% 
                       summarize(mean_press = mean(press_mean),
                                 sd_press = sd(press_mean),
                                 min_press = min(press_min),
                                 max_press = max(press_max),
                                 mean_resp = mean(resp),
                                 min_resp = min(resp),
                                 max_resp = max(resp),
                                 sd_resp = sd(resp),
                                 mean_temp = mean(temp_mean),
                                 min_temp = min(temp_min),
                                 max_temp = max(temp_max),
                                 sd_temp = sd(temp_mean),
                                 mean_speed = mean(speedBL),
                                 min_speed = min(speedBL_min),
                                 max_speed = max(speedBL_max),
                                 sd_speed = sd(speedBL)
                                 ) 
  

flextable(summ_by_ID %>%
            group_by(ntemp, npress) %>% 
            summarise(press_mean = round(mean(mean_press), 1),
                      press_sd = round(sd(mean_press), 1),
                      press_min = round(min(min_press),1),
                      press_max = round(max(max_press),1),
                      resp_mean = round(mean(mean_resp), 1),
                      resp_sd = round(sd(mean_resp), 1),
                      resp_min = round(min(mean_resp), 1),
                      resp_max = round(max(mean_resp), 1),
                      temp_mean = round(mean(mean_temp), 1),
                      temp_sd = round(sd(mean_temp), 1),
                      temp_min = round(min(mean_temp), 1),
                      temp_max = round(max(mean_temp), 1),
                      speed_mean = round(mean(mean_speed), 2),
                      speed_sd = round(sd(mean_speed), 2),
                      speed_min = round(min(mean_speed), 2),
                      speed_max = round(max(mean_speed), 2)) %>% 
            mutate(press_range = paste(press_min, press_max, sep = "-"),
                   resp_mean = paste(resp_mean, resp_sd, sep ="\u00B1"), 
                   resp_range = paste(resp_min, resp_max, sep = "-"),
                   temp_mean = paste(temp_mean, temp_sd, sep ="\u00B1"), 
                   temp_range = paste(temp_min, temp_max, sep = "-"),
                   speed_mean = paste(speed_mean, speed_sd, sep ="\u00B1"), 
                   speed_range = paste(speed_min, speed_max, sep = "-")) %>%
            ungroup() %>% 
            select(press_range, temp_range, speed_range, resp_mean) %>%
            rename("Observed pressure range (bar)" = press_range,
                 "Mean oxygen consupmtion rate (mg/kg/h)" = resp_mean,
                 "Observed temperature range (°C)" = temp_range,
                 "Observed velocity range (BL/s)" = speed_range)) %>% 
  autofit()
                     


```

\
\
\
**Tab 2** Short summary of the model statistics for fixed effects. Note, that value gives the change in the log of O2-consuption rate per change in one unit of the predictor (main effects refer to y-axis intercept with all predictors = 0).

```{r Tab 2}

intervals <- intervals(lme.plot.center)
intervals <- data.frame(intervals$fixed)
intervals <- intervals %>%
            rename(Lower_confidence_limit = lower,
                   Upper_confidence_limit = upper)
result <- as.data.frame(coef(summary(lme.plot.center)))
result$p <- result$'p-value'
result$Parameter <- rownames(result)
intervals$rownames <- rownames(intervals)


flextable(result %>%
  mutate(p = ifelse(p <= 0.001, "<0.001" , round(p, 3))) %>%
  left_join(intervals, by = c("Parameter" = "rownames")) %>%
  mutate("Lower conf. limit" = round(Lower_confidence_limit, 3),
         "Upper conf. limit" = round(Upper_confidence_limit, 3),
         Value = round(Value, 3)) %>%
  select(c('Parameter', 'Value', 'Lower conf. limit', 'Upper conf. limit', 'p')) %>% 
  mutate(Parameter = if_else(Parameter == "(Intercept)", "Intercept", 
                             if_else(Parameter == "temp_mean_center", "Temperature (°C)",
                             if_else(Parameter == "press_mean_center", "Pressure (bar)",
                             if_else(Parameter == "speedBL_center", "Velocity (BL/s)", 
                             if_else(Parameter == "temp_mean_center:press_mean_center", "Temp/Press Interaction", "NA"))))))) %>%
  autofit() 
 

```

\
\
\
\
\
\
\

## Graphs for the manuscript



\
\
\

```{r Fig 2, dpi = 300}

#ideal_summary <- ideal %>%
#  mutate(group = paste(temp_mean, press_mean, sep = "_")) %>% 
#  group_by(group) %>% 
#  summarize(mean = mean(resp),
#            sd = sd(resp))



#fig2a <-  tag_facet2(ggplot(ideal_1, aes(x=temp_mean, y=resp)) +
#  #geom_point(aes(colour = temp_mean)), size=1)+ 
#  theme_bw()+
#  stat_smooth(data = ideal_1, aes(x= temp_mean, y=resp, color = ID), se = F, size = 0.5, alpha = 0.75, geom="line", linetype = #"solid")+
#  stat_smooth(data = ideal_1, aes(x= temp_mean, y=resp_glob), se = F, size = 1, color = "black", alpha = 0.7, geom="line")+
#  geom_point(data = summary, aes(x= temp_mean, y = resp, color = ID), alpha = 0.5)+
#  #scale_colour_scico(palette = "batlow")+
#  scale_colour_discrete(guide = "none")+
#  #scale_linetype_manual(values=c(1,2), guide = "none") + 
#  labs(linetype = "pressure (bar)", x = "temp (°C)", y = expression(O[2]-consumption ~(mg/kg %*% h^-1)))+
#  facet_grid(run ~ tunnel, labeller = labeller(tunnel = tunnel.names, run = run.names))+
#  theme(legend.position = c(.05, .95),
#        legend.justification = c("left", "top"),
#        legend.margin = margin(6, 6, 6, 6),
#        legend.box.background = element_rect(color="black", size=1))+
#  theme(text = element_text(size = 10, family = "sans")))


#fig2b <-  tag_facet2(ggplot(ideal, aes(x=temp_mean, y=resp)) +
#  theme_bw()+
#  geom_point(data = summary, aes(x= temp_mean, y = resp, color = ID, shape = as.factor(npress)), alpha = 0.5)+
#  scale_shape_manual(values = c(1, 16), guide = "none")+
#  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 1))), color = ID), se = F, size = 0.4, geom="line", linetype = "solid")+
#  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 8))), color = ID), se = F, size = 0.4, geom="line", linetype = "dashed")+
#  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 1), level = "0"))), se = F, size = 0.4, color = "black", geom="line", linetype = "solid")+
#  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 8), level = "0"))), se = F, size = 0.4, color = "black", geom="line", linetype = "dashed")+
#  scale_colour_discrete(guide = "none")+
#  labs(linetype = "pressure (bar)", x = "temp (°C)", y = expression(O[2]-consumption ~(mg/kg %*% h^-1)))+
#  facet_grid(run ~ tunnel, labeller = labeller(tunnel = tunnel.names, run = run.names))+
#  theme(legend.position = c(.05, .95),
#        legend.justification = c("left", "top"),
#        legend.margin = margin(6, 6, 6, 6),
#        legend.box.background = element_rect(color="black", size=1))+
#  theme(text = element_text(size = 10, family = "sans")))
#
#fig2c_old <-  tag_facet2(ggplot(ideal, aes(x=temp_mean, y=resp)) +
#  theme_bw()+
#  geom_point(data = summary, aes(x= temp_mean, y = resp, color = speedBL, shape = as.factor(npress)), alpha = 0.5)+
#  scale_shape_manual(values = c(1, 16), guide = "none")+
#  scale_colour_scico(palette = "batlow")+
#  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 1)))), se = F, size = 0.4, geom="line", linetype = "solid", color = "red")+
#  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 8)))), se = F, size = 0.4, geom="line", linetype = "dashed", color = "red")+
#  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 1), level = "0"))), se = F, size = 0.4, color = "black", geom="line", linetype = "solid")+
#  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=exp(predict(lme.plot, newdata = ideal %>% filter(press_mean #== 8), level = "0"))), se = F, size = 0.4, color = "black", geom="line", linetype = "dashed")+
#  #scale_linetype_manual(values=c(1,2), guide = "none") + 
#  labs(linetype = "pressure (bar)", x = "temp (°C)", y = expression(O[2]-consumption ~(mg/kg %*% h^-1)))+
#  facet_grid(run ~ tunnel, labeller = labeller(tunnel = tunnel.names, run = run.names))+
#  theme(legend.key.size = unit(3, "cm"),
#      legend.key.width = unit(0.3, "cm"),
#      legend.title = element_blank(),
#      text = element_text(size = 10, family = "sans")))

fig2c <-  tag_facet2(
  ggplot(summary, aes(x=temp_mean, y=log_resp)) +
  theme_bw()+
  geom_point(aes(x= temp_mean, y = log_resp, color = speedBL, shape = as.factor(npress)), alpha = 0.5)+
  scale_shape_manual(values = c(16, 1), guide = "none")+
  scale_colour_scico(palette = "batlow")+
  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=predict(lme.plot, newdata = ideal %>% filter(press_mean == 1))), se = F, size = 0.4, geom="line", linetype = "solid", color = "red")+
  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=predict(lme.plot, newdata = ideal %>% filter(press_mean == 8))), se = F, size = 0.4, geom="line", linetype = "dashed", color = "red")+
  stat_smooth(data = ideal %>% filter(press_mean == 1), aes(x= temp_mean, y=predict(lme.plot, newdata = ideal %>% filter(press_mean == 1), level = "0")), se = F, size = 0.4, color = "black", geom="line", linetype = "solid")+
  stat_smooth(data = ideal %>% filter(press_mean == 8), aes(x= temp_mean, y=predict(lme.plot, newdata = ideal %>% filter(press_mean == 8), level = "0")), se = F, size = 0.4, color = "black", geom="line", linetype = "dashed")+
  #scale_linetype_manual(values=c(1,2), guide = "none") + 
  labs(linetype = "pressure (bar)", x = "temp (°C)", y = expression(ln(O[2]-consumption ~(mg/kg %*% h^-1))))+
  facet_grid(run ~ tunnel, labeller = labeller(tunnel = tunnel.names, run = run.names))+
  guides(scale="none"), open ="", close ="", hjust = -1.2) +
  theme(legend.key.size = unit(3, "cm"),
      legend.key.width = unit(0.3, "cm"),
      legend.title = element_blank(),
      text = element_text(size = 10, family = "sans"))


#fig2d <-  ggplot(ideal_1, aes(x=temp_mean, y=resp)) +
#  #geom_point(aes(colour = temp_mean)), size=1)+ 
#  theme_bw()+
#  stat_smooth(data = ideal_1, aes(x= temp_mean, y=resp, group = ID, color = startmass), se = F, size = 0.5, alpha = 0.75, geom="line", #linetype = "dashed")+
#  stat_smooth(data = ideal_1, aes(x= temp_mean, y=resp_glob), se = F, size = 1, color = "black", alpha = 0.7, geom="line")+
#  scale_colour_scico(palette = "batlow")+
#  #scale_colour_discrete(guide = "none")+
#  #scale_linetype_manual(values=c(1,2), guide = "none") + 
#  labs(linetype = "pressure (bar)", x = "temp (°C)", y = expression(O[2]-consumption ~(mg/kg %*% h^-1)))+
#  theme(legend.position = c(.05, .95),
#        legend.justification = c("left", "top"),
#        legend.margin = margin(6, 6, 6, 6),
#        legend.box.background = element_rect(color="black", size=1))+
#  theme(text = element_text(size = 10, family = "sans"))

#ggsave("Fig2a.pdf", fig2a, device = "pdf", height = 7, width = 8, dpi = 300)
#ggsave("Fig2b.pdf", fig2b, device = "pdf", height = 7, width = 8, dpi = 300)
ggsave("Fig1.pdf", fig2c, device = "pdf", height = 7, width = 8, dpi = 300)
#ggsave("Fig2d.pdf", fig2d, device = "pdf", height = 7, width = 8, dpi = 300)

#fig2a
#fig2b
fig2c
#fig2d

```

**Fig 1** Observed vs predicted oxygen consumption rates at different temperatures, pressures (solid circles/lines: ~1bar, open circles/dashed line: ~8bar) and velocities (in BL/s). Note, that ID's A and B were removed from statistical analyses due to a
technical issue with the pumps during flushing and therefore no predictions are presented.
\
\
\
\
\
\

## Graphs/tabs for supplementary

**Tab S1** Summary of experimental parameters and oxygen consumption grouped by individual, nominal temperature and pressure. Note, that ID's A & B were removed from analysis completely and for ID F, two measurements were removed for analysis. 

```{r}


calc_n <- summary %>% mutate(indicator = paste(ID, ntemp, npress, sep = "_")) %>% 
                      group_by(indicator) %>% 
                      summarize(n = length(estimate_s)) 

summ_by_ID2 <- summary %>% 
            group_by(ID, ntemp, npress) %>% 
                       summarize(mean_press = mean(press_mean),
                                 sd_press = sd(press_mean),
                                 min_press = min(press_min),
                                 max_press = max(press_max),
                                 mean_resp = mean(resp),
                                 min_resp = min(resp),
                                 max_resp = max(resp),
                                 sd_resp = sd(resp),
                                 mean_temp = mean(temp_mean),
                                 min_temp = min(temp_min),
                                 max_temp = max(temp_max),
                                 sd_temp = sd(temp_mean),
                                 mean_speed = mean(speedBL),
                                 min_speed = min(speedBL_min),
                                 max_speed = max(speedBL_max),
                                 sd_speed = sd(speedBL),
                                 time = as.numeric(abs(difftime(min(datetime_min), max(datetime_max), units = "days")))
                       )
                                 

addresp <- summ_by_ID2 %>% 
  ungroup() %>% 
  mutate(indicator = paste(ID, ntemp, npress, sep = "_")) %>% 
  select(-ntemp, -npress, -ID) %>%
  mutate_at(1:16, round, 2) %>%
  #mutate_at(13:16, round, 2) %>%
  mutate(p = paste(mean_press, sd_press, sep = " \u00B1 "),
         p_range = paste(min_press, max_press, sep = " - "),
         t = paste(mean_temp, sd_temp, sep = " \u00B1 "),
         t_range = paste(min_temp, max_temp, sep = " - "),
         o = paste(mean_resp, sd_resp, sep = " \u00B1 "),
         o_range = paste(min_resp, max_resp, sep = " - "),
         s = paste(mean_speed, sd_speed, sep = " \u00B1 "),
         s_range = paste(min_speed, max_speed, sep = " - "),
         time = round(time, 1)) %>% 
  select(indicator,p, p_range, t, t_range, o, o_range, s, s_range, time) %>% 
  rename("Mean observed pressure (bar)" = p,
         "Observed pressure range (bar)" = p_range,
         "Mean observed temperature (°C)" = t,
         "Observed temperature range (°C)" = t_range,
         "Mean oxygen consumption (mg/kg/h)" = o,
         "Oxygen consumption range (mg/kg/h)" = o_range,
         "Mean observed velocity (BL/s)" = s,
         "Observed velocity range (BL/s)" = s_range,
         "time swam (days)" = time)
         
flextable(biomasses %>%
            select(ID, ID_old) %>% 
            left_join(treattime, by = c("ID_old" = "ID")) %>%
            mutate(ntemp = replace(ntemp, ntemp == "15", 15.5),
                   indicator = paste(ID_old, ntemp, npress, sep = "_")) %>%
            left_join(calc_n, by = "indicator") %>%
            left_join(addresp, by = "indicator") %>% 
            mutate(time_d = round(time_d, 1)) %>%
            rename("swim time (d)" = time_d) %>%
            select(ID, "Mean observed temperature (°C)", "Observed temperature range (°C)", "Mean observed pressure (bar)", "Observed pressure range (bar)", "Mean observed velocity (BL/s)", "Observed velocity range (BL/s)", "time swam (days)", n, "Mean oxygen consumption (mg/kg/h)", "Oxygen consumption range (mg/kg/h)")) %>%
          colformat_num(big.mark   = "") %>%
  autofit()

```

**Tab S2** Summary of candidate models with respective selection criteria

```{r Tab S1}

round_df <- function(x, digits) {
    # round all numeric variables
    # x: data frame 
    # digits: number of digits to round
    numeric_columns <- sapply(x, mode) == 'numeric'
    x[numeric_columns] <-  round(x[numeric_columns], digits)
    x
}

remotes::install_github("davidgohel/flextable")
library(flextable)

AIC_scores <- select.lme %>%
         mutate_if(is.numeric, round, 3) %>% 
         replace(is.na(.), "") %>% 
         slice(1:5)
                
AIC_scores <- flextable(AIC_scores) 

FitFlextableToPage <- function(ft, pgwidth = 6){

  ft_out <- ft %>% autofit()

  ft_out <- width(ft_out, width = dim(ft_out)$widths*pgwidth /(flextable_dim(ft_out)$widths))
  return(ft_out)
}

FitFlextableToPage(AIC_scores)
 
```

\
\
\
\
```{r Fig S1, dpi = 300}

fig1 <- tag_facet2(
  ggplot(summary, aes(x=ID_time, y=log(resp))) +
  geom_point(data = summary, aes(col = temp_mean, shape = as.factor(npress)), size=0.9, alpha = 0.5) +
  theme_bw()+
  scale_shape_manual(values = c(16, 1), guide = "none")+
  scale_colour_scico(palette = "batlow")+ # ,name = "temp (°C)" - this adds title for the specific legend
  geom_line(data= summary_model, aes(x= ID_time, y=predict(lme.plot, newdata = summary_model, level = 0)), linewidth = 0.4)+
  geom_line(data = summary_model, aes(x= ID_time, y=predict(lme.plot, newdata = summary_model), group = ID), linetype = "dashed", size = 0.4)+
  labs(x = expression(time~(h)), y = expression(ln~O[2]-consumption ~(mg/kg %*% h^-1)))+
  facet_grid(run ~ tunnel, labeller = labeller(tunnel = tunnel.names, run = run.names))+
  guides(scale="none"), open ="", close ="", hjust = -1.2) +
  theme(legend.key.size = unit(3, "cm"),
      legend.key.width = unit(0.3, "cm"),
      legend.title = element_blank(),
      text = element_text(size = 10, family = "sans"))

fig1

#add plot with temp for qual check
 # ggplot(summary, aes(x=datetime_min, y=temp_mean)) +
 #   theme_bw() +
 #   geom_point(aes(col = tunnel, shape = as.factor(npress)), alpha = 0.5)+
 #   scale_shape_manual(values = c(16, 1), guide = "none") +
 #   scale_color_manual(values = scico(3, begin = 0.3, palette = "batlow"))+
 #   labs(color='Tunnel')+
 #   xlab("Date")+
 #   ylab("Mass (g)")
    
  
#add graph with masses for quality check
 #   ggplot(summary %>% mutate(Letter_ID = toupper(Letter_ID)), aes(x=datetime_min, y=mass_now)) +
 #   theme_bw() +
 #   geom_point(aes(col = Letter_ID, shape = as.factor(npress)), alpha = 0.5)+
 #   scale_shape_manual(values = c(16, 1), guide = "none") +
 #   scale_color_manual(values = scico(9, begin = 0.2, palette = "batlow"))+
 #   labs(color='ID')+
 #   xlab("Date")+
 #   ylab("Mass (g)")

ggsave("FigS1.png", fig1, device = "png", height = 7, width = 8, dpi = 300)

```
**Fig S1** Oxygen consumption rate per individual (a-i) over time different at temperatures (color scale) at 1 bar (open circles) and 8 bar (closed circles). Modeled predictions are displayed on the population level (solid line) and for the individual (dashed line). Colour indicates temp in °C. Note, that ID's A and B were removed from statistical analyses due to a technical issue with the pumps\
during flushing and therefore no predictions are presented.
\
\
\
\



```{r Fig S2, dpi = 600}

ggplot(summary %>% mutate(ntemp = ifelse(ntemp == 12, 13,
                                         ifelse(ntemp==15.5, 16, 
                                                ifelse(ntemp==19,19,NA)))), aes(x=speedBL, y=resp)) +
    geom_point(size=0.2)+ 
    theme_bw()+
    theme(text = element_text(size = 8, family = "sans")) +
    xlim(0.4, 0.8)+
    facet_nested(ntemp + npress ~ Letter_ID )+
  labs(linetype = "pressure (bar)", x = "flow rate (body length/s)", y = expression(O[2]-consumption ~(mg/kg %*% h^-1)))+
  theme(axis.text.x = element_text(angle = 45, size = 6))+
  theme(axis.text.y = element_text(size = 8))+
  geom_smooth(method='lm', formula= y~x, se = F, size = 0.7)
```

**Fig S2** Effect of swimming speed on Oxygen consumption rate per Temperature and pressure for each individual
